Allow missing dependencies while installing the base app
flatpak-bundler can be used to just shuffle files into a flatpak
without actually building anything. This is particularly useful for
building electron like apps which are just dealing with precompiled
binaries.

In that case we don't really need the sdk or platform around, and
should support them being missing. We can do this by passing the
--no-deps flag command while autoinstalling the base app.

https://phabricator.endlessm.com/T14918